requirements.txt
Django==1.8
wsgiref
django-concurrency==1.3.2
djangorestframework==3.5.3
drf-nested-routers==0.11.1
pep8
